test: move the pipeline wiring version tests to their own file
Keeps test_models.py under the file-length threshold CodeScene holds it to.
